Module Vgr.Manual

Manual render destinations.

Warning. Only use with renderers with `Manual destinations.

val dst : renderer -> bytes -> int -> int -> unit

dst r s j l provides r with l bytes to write, starting at j in s. This byte range is written by calls to render until `Partial is returned. Use dst_rem to know the remaining number of non-written free bytes in s.

val dst_rem : renderer -> int

dst_rem r is the remaining number of non-written, free bytes in the last buffer provided with dst.