Module Private.P

Paths helpers.

val of_data : Data.path -> P.t

of_data d is the path from the internal representation d.

val earc_params : Gg.p2 -> large:bool -> cw:bool -> float -> Gg.v2 -> Gg.p2 -> (Gg.p2 * Gg.m2 * float * float) option

earc_params p large cw angle r p' is Some (c, m, a, a') with c the center of the ellipse, m a transform matrix mapping the unit circle to the ellipse, a and a' the angle on the unit circle corresponding to the first and last point of the arc. None is returned if the parameters do not define a valid arc.

val miter_limit : P.outline -> float

miter_limit o is the miter limit corresponding to o.miter_angle.